+++ lib/src/channels/channel_client.dart
@@ -8,7 +8,6 @@ import '../reverse_engineering/pages/watch_page.dart';
 import '../reverse_engineering/youtube_http_client.dart';
 import '../videos/video.dart';
 import '../videos/video_id.dart';
-import 'channel_handle.dart';
 import 'channels.dart';
 
 /// Queries related to YouTube channels.